+++ lib/openmaize/config.ex
@@ -11,9 +11,6 @@ defmodule Openmaize.Config do
   | db_module          | module  | Openmaize.DB   |
   | hash_name          | atom    | :password_hash |
   | crypto_mod         | atom    | :bcrypt  |
-  | token_alg          | atom    | :sha512  |
-  | token_validity     | int     | 120 (minutes)  |
-  | keyrotate_days     | int     | 28       |
   | password_strength  | keyword list | []  |
 
   The values for user_model and repo should be module names.
@@ -24,7 +21,8 @@ defmodule Openmaize.Config do
   ## Examples
 
   The simplest way to change the default values would be to add
-  the following to the `config.exs` file in your project.
+  an `openmaize` entry to the `config.exs` file in your project,
+  like the following example.
 
       config :openmaize,
         user_model: Coolapp.User,
@@ -32,9 +30,6 @@ defmodule Openmaize.Config do
         db_module: Coolapp.DB,
         hash_name: :encrypted_password,
         crypto_mod: :pbkdf2,
-        token_alg: :sha256,
-        token_validity: 60,
-        keyrotate_days: 7,
         password_strength: [min_length: 12]
 
   """
@@ -86,37 +81,6 @@ defmodule Openmaize.Config do
   end
   defp crypto_mod do
     Application.get_env(:openmaize, :crypto_mod, :bcrypt)
-  end
-
-  @doc """
-  The algorithm used to sign the token.
-
-  The default value is :sha512, and :sha256 is also supported.
-  """
-  def get_token_alg do
-    case token_alg do
-      :sha256 -> {"HS256", :sha256}
-      _ -> {"HS512", :sha512}
-    end
-  end
-  defp token_alg do
-    Application.get_env(:openmaize, :token_alg, :sha512)
-  end
-
-  @doc """
-  The length of time after which a JSON Web Token expires.
-
-  The default length of time is 120 minutes (2 hours).
-  """
-  def token_validity do
-    Application.get_env(:openmaize, :token_validity, 120)
-  end
-
-  @doc """
-  The number of days after which the JWT signing keys will be rotated.
-  """
-  def keyrotate_days do
-    Application.get_env(:openmaize, :keyrotate_days, 28)
   end
 
   @doc """