+++ lib/openmaize/onetime_pass.ex
@@ -1,5 +1,12 @@
 defmodule Openmaize.OnetimePass do
   @moduledoc """
+  Module to handle one-time passwords for use in two factor authentication.
+
+  There is one option
+
+  * add_jwt - the function used to add the JSON Web Token to the response
+    * the default is `&OpenmaizeJWT.Plug.add_token/3`
+
   """
 
   import Plug.Conn
@@ -9,32 +16,38 @@ defmodule Openmaize.OnetimePass do
   @behaviour Plug
 
   def init(opts) do
-    Keyword.pop opts, :unique_id, :username
+    Keyword.pop opts, :add_jwt, &OpenmaizeJWT.Plug.add_token/3
   end
 
   @doc """
+  Handle the one-time password POST request.
+
+  If the one-time password check is successful, a JSON Web Token will be added
+  to the conn, either in a cookie or in the body of the response. The conn
+  is then returned.
   """
-  def call(%Plug.Conn{params: %{"user" => user_params}} = conn, {uniq, opts}) do
-    case Map.get(user_params, to_string(uniq)) do
-      nil -> handle_auth(false, conn)
-      user_id ->
-        Config.db_module.find_user(user_id, uniq)
-        |> check_key(user_params, opts)
-        |> handle_auth(conn)
-    end
+  def call(%Plug.Conn{params: %{"user" => user_params}} = conn, {add_jwt, opts}) do
+    {storage, uniq, id} = get_params(user_params)
+    Config.db_module.find_user_byid(id)
+    |> check_key(user_params, opts)
+    |> handle_auth(conn, {storage, uniq, add_jwt})
   end
 
-  def check_key(user, %{"hotp" => hotp}, opts) do
-    Otp.check_hotp(hotp, user.otp_secret, opts)
-  end
-  def check_key(user, %{"totp" => totp}, opts) do
-    Otp.check_totp(totp, user.otp_secret, opts)
+  defp get_params(%{"storage" => storage, "uniq" => uniq, "id" => id}) do
+    {String.to_atom(storage), String.to_atom(uniq), id}
   end
 
-  def handle_auth(false, conn) do
+  defp check_key(user, %{"hotp" => hotp}, opts) do
+    {user, Otp.check_hotp(hotp, user.otp_secret, opts)}
+  end
+  defp check_key(user, %{"totp" => totp}, opts) do
+    {user, Otp.check_totp(totp, user.otp_secret, opts)}
+  end
+
+  defp handle_auth({_, false}, conn, _opts) do
     put_private(conn, :openmaize_error, "Invalid credentials")
   end
-  def handle_auth(last, conn) do
-    put_private(conn, :openmaize_info, last)
+  defp handle_auth({user, last}, conn, {storage, uniq, add_jwt}) do
+    conn |> put_private(:openmaize_info, last) |> add_jwt.(user, {storage, uniq})
   end
 end