+++ test/openmaize/onetime_pass_test.exs
@@ -10,31 +10,39 @@ defmodule Openmaize.OnetimePassTest do
     trunc((megasecs * 1000000 + secs) / 30)
   end
 
-  def call(uniq, name, otp_type, otp, opts) do
-    conn(:post, "/twofactor",
-         %{"user" => %{uniq => name, otp_type => otp}})
-    |> OnetimePass.call(opts)
+  def call(user, opts) do
+    conn(:post, "/twofactor", %{"user" => user}) |> OnetimePass.call(opts)
   end
 
   test "check hotp with default options" do
-    conn = call("username", "brian", "hotp", "816065", {:username, []})
+    user = %{"hotp" => "816065", "storage" => "cookie", "uniq" => "username", "id" => "5"}
+    conn = call(user, {&OpenmaizeJWT.Plug.add_token/3, []})
+    assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info] == 2
     refute conn.private[:openmaize_error]
-    conn = call("username", "brian", "hotp", "816066", {:username, []})
+    fail = %{"hotp" => "816066", "storage" => "cookie", "uniq" => "username", "id" => "5"}
+    conn = call(fail, {&OpenmaizeJWT.Plug.add_token/3, []})
+    refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
   end
 
   test "check hotp with last option" do
-    conn = call("username", "brian", "hotp", "088239", {:username, [last: 18]})
+    user = %{"hotp" => "088239", "storage" => "cookie", "uniq" => "username", "id" => "5"}
+    conn = call(user, {&OpenmaizeJWT.Plug.add_token/3, [last: 18]})
+    assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info] == 19
     refute conn.private[:openmaize_error]
-    conn = call("username", "brian", "hotp", "088238", {:username, [last: 18]})
+    fail = %{"hotp" => "088238", "storage" => "cookie", "uniq" => "username", "id" => "5"}
+    conn = call(fail, {&OpenmaizeJWT.Plug.add_token/3, [last: 18]})
+    refute conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_error]
   end
 
   test "check totp with default options" do
     token = Otp.gen_totp("MFRGGZDFMZTWQ2LK")
-    conn = call("email", "brian@mail.com", "totp", token, {:email, []})
+    user = %{"totp" => token, "storage" => "cookie", "uniq" => "email", "id" => "5"}
+    conn = call(user, {&OpenmaizeJWT.Plug.add_token/3, []})
+    assert conn.resp_cookies["access_token"]
     assert conn.private[:openmaize_info]
     refute conn.private[:openmaize_error]
   end